+++ docs/start/changelog.rst
@@ -6,6 +6,8 @@ ChangeLog
 0.24.5 - TBD
     * Fixed a mistake in the product registry
       (``LCM3_MINI2_WARM_WHITE`` should be ``WARM_TO_WHITE``)
+    * Fixed multi options for the setting of strip Color Zones using legacy
+      messages.
 
 0.24.4 - 6 January 2020
     * Made it possible to override target options from the command line.